postprocess-service/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
kafka-python==2.0.2
boto3==1.34.0
python-dotenv==1.0.0
Pillow==10.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0 